Use signed wind angle convention [-180,180]

+++ sketch/RUDDER.h
@@ -1,17 +1,19 @@
 #ifndef RUDDER_H
 #define RUDDER_H
 
+#include <Arduino.h>
+
 class RUDDER {
 public:
     RUDDER();
     void init();
     void update();
     void set_rudder_angle(int angle);
-    float get_rudder_angle();
+    int16_t get_rudder_angle();
 private:
-    float rudder_angle;
+    int16_t rudder_angle;
     int ADC_rudder_value;
-    float k;
+    int16_t k;
     void test_rudder_servo();
 };
